spi-tty-ds [backport]: add SPI TTY driver from kernel 3.8
This is a simple TTY driver that use the SPI bus to communicate. We use this
driver to communicate with the MCU on the Linino's boards. These boards have
their own SPI driver in order to make the link active.

Typically the communication with an MCU is done with a real serial interface,
so there are dozen of high level libraries that assume that this TTY link is
active. In order to avoid to fix libraries before use them in Linino, we
decided to hide the SPI bus behind the TTY interface. This way, no work is
needed on those libraries.

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>