[BUG] .config: remove yun-conf package
The yun-conf package has a compile problem. The problem is due to a
modification in the OpenWRT building system.

The building system forces the building of the 'base-files' package to
be the last, even if there is a dependency.

The 'yun-conf' packade needs 'base-files' files to be succesfully
installed. This is not possible any more.

The package is removed from the default configuration and marked as
BROKEN.

https://dev.openwrt.org/ticket/10257

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>